requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pycodestyle==2.8.0
pyTelegramBotAPI==4.5.1
python-environ==0.4.54
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
